+++ CHANGELOG.md
@@ -10,6 +10,8 @@ PowerModels.jl Change Log
 - Fixed a mathematical bug when swapping the orientation of a transformer
 - Added support for parsing of PTI raw files into a Dict
 - Updated branch mathematical model and Matpower parser to support asymmetrical line charging
+- Updated KCL constraint models to account for multiple loads and shunts per bus
+- Separated loads and shunts from buses in PowerModels data structure
 
 ### v0.5.1
 - Added support for convex piecewise linear cost functions